btrfs-progs: tests: check for partscan support in misc/006-partitioned-loopdev
The travis CI does not have losetup with --partscan, skip the check in
that case so we can still run the mkfs testsuite.

+++ tests/mkfs-tests/006-partitioned-loopdev/test.sh
@@ -3,6 +3,11 @@
 
 source $TOP/tests/common
 
+if ! losetup --help | grep -q 'partscan'; then
+	_not_run "losetup --partscan not available"
+	exit 0
+fi
+
 check_prereq mkfs.btrfs
 
 setup_root_helper
